Skip re-registering subscriptions that are not about to expire
