+++ app/src/main/kotlin/at/bitfire/davdroid/push/PushRegistrationWorkerManager.kt
@@ -11,7 +11,12 @@ import androidx.work.NetworkType
 import androidx.work.PeriodicWorkRequest
 import androidx.work.WorkManager
 import at.bitfire.davdroid.repository.DavCollectionRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import dagger.multibindings.IntoSet
 import kotlinx.coroutines.runBlocking
 import java.util.concurrent.TimeUnit
 import java.util.logging.Logger
@@ -42,7 +47,7 @@ class PushRegistrationWorkerManager @Inject constructor(
             workManager.enqueueUniquePeriodicWork(
                 UNIQUE_WORK_NAME,
                 ExistingPeriodicWorkPolicy.CANCEL_AND_REENQUEUE,
-                PeriodicWorkRequest.Builder(PushRegistrationWorker::class, 1, TimeUnit.DAYS)
+                PeriodicWorkRequest.Builder(PushRegistrationWorker::class, INTERVAL_DAYS, TimeUnit.DAYS)
                     .setInitialDelay(5, TimeUnit.SECONDS)
                     .setConstraints(
                         Constraints.Builder()
@@ -60,6 +65,32 @@ class PushRegistrationWorkerManager @Inject constructor(
 
     companion object {
         private const val UNIQUE_WORK_NAME = "push-registration"
+        const val INTERVAL_DAYS = 1L
     }
 
+
+    /**
+     * Listener that enqueues a push registration worker when the collection list changes.
+     */
+    class CollectionsListener @Inject constructor(
+        @ApplicationContext val context: Context,
+        val workerManager: PushRegistrationWorkerManager
+    ): DavCollectionRepository.OnChangeListener {
+
+        override fun onCollectionsChanged() {
+            workerManager.updatePeriodicWorker()
+        }
+
+    }
+
+    /**
+     * Hilt module that registers [CollectionsListener] in [DavCollectionRepository].
+     */
+    @Module
+    @InstallIn(SingletonComponent::class)
+    interface PushRegistrationWorkerModule {
+        @Binds
+        @IntoSet
+        fun listener(impl: CollectionsListener): DavCollectionRepository.OnChangeListener
+    }
 }
